James Moran CGP600 AE2
Updates, with the report now showing a render-view of the Player's ship
side-by-side with the reference image, along with the ship's static-mesh
shown in the default game scene. I have also removed an asset from the
Max project (as it is not required any-more, as I was able to resolve
rendering issues in-game, with Visual Studio 2015).