AGP 1.0.20
With the revamped box to have been used as a splash screen, as well as
an updated report to reflect on the changes made to the game, so that
the Player can win/lose (even though they are not informed about how
either of these states are possible). I have also updated the Mobile
Applications project to include the google-play-services_lib, as well as
submitting a .apk file to the shared Google Drive Folder, for Anthony to
use for testing our game.