FMP Logbook 1.0.36
As well as the calculations for the Dispersion Coefficient values, in an
Excel spreadsheet, along with the implementation of the calcuations for
the other Coefficient values. These calculations turn out as expected
(with no errors). Determine a means to compare these values, in order to
generate a balanced FPS level.